+++ common/model/src/main/java/is/codion/common/model/preferences/PreferencesPath.java
@@ -97,18 +97,18 @@ final class PreferencesPath {
 	private static Path linuxPath(String filename) {
 		String xdgConfigHome = System.getenv("XDG_CONFIG_HOME");
 		if (xdgConfigHome != null) {
-			return Paths.get(xdgConfigHome, CODION_DIR.toLowerCase(), filename);
+			return Paths.get(xdgConfigHome, CODION_DIR.toLowerCase(Locale.ROOT), filename);
 		}
 
 		String home = System.getProperty(USER_HOME);
 
-		return Paths.get(home, ".config", CODION_DIR.toLowerCase(), filename);
+		return Paths.get(home, ".config", CODION_DIR.toLowerCase(Locale.ROOT), filename);
 	}
 
 	private static Path defaultPath(String filename) {
 		String home = System.getProperty(USER_HOME);
 
-		return Paths.get(home, "." + CODION_DIR.toLowerCase(), filename);
+		return Paths.get(home, "." + CODION_DIR.toLowerCase(Locale.ROOT), filename);
 	}
 
 	private static void validateFilename(String filename) {